+++ src/components/Menu/MenuItemGroup.tsx
@@ -1,7 +1,10 @@
 import { styled } from '@mui/material';
 
-export const EnteMenuItemGroup = styled('div')(
+export const MenuItemGroup = styled('div')(
     ({ theme }) => `
+    & > .MuiMenuItem-root{
+        border-radius: 8px; 
+    }
     & > .MuiMenuItem-root:not(:last-of-type) {
         border-bottom-left-radius: 0;
         border-bottom-right-radius: 0;
@@ -11,6 +14,6 @@ export const EnteMenuItemGroup = styled('div')(
         border-top-right-radius: 0;
     }
     background-color: ${theme.colors.background.elevated2};
-    border-radius: 4px;
+    border-radius: 8px;
 `
 );
